requirements.txt
pandas
matplotlib
seaborn
torchvision
Pillow
plantcv
flake8